Changes to make it compile on Microsoft Visual C++ 2008

+++ src/pwancmdlineparser.h
@@ -36,7 +36,7 @@ namespace pwan
        private:
             std::vector<optionsReturn> setOptions;
             std::vector<optBlob> allowedOptions;
-            int defaultOpt;
+            size_t defaultOpt;
     };
 }
